compute hash is calling recompute always
When it does the put of a new hash it's not using the same key as it uses
when it fetches, and the get function only returns the hash value not the full
key pair. This is obviously not the best fix, but works for now. Not sure how often
this is used in catenary, but it might add up. I'll dig into this further and add some
tests.